Improved python implemntation Changed from coordinate based input to city/state input (still only works for US)
